Openalyx switch, README updates, moved load_regressors
